Add DruidNet logo in the top left corner when the plant is in the database
This commit adds the DruidNet logo as a badge to the top-left corner of the main image on the plant identification success screen (`PlantInfoDruidNet`). This visually indicates that the identified plant is present in the DruidNet database.

Additionally, a new `@Preview` function, `PlantInfoDruidNetPreview`, has been created to facilitate the isolated testing and development of this component.