feat: Compress images before uploading for identification
This commit introduces an image compression step before uploading a plant picture for identification. This is done to reduce upload size and improve performance.

A new utility function, `compressImage`, has been created in `ImageUtils.kt`. This function takes an image `Uri`, scales it down to a quarter of its size, and compresses it into a temporary JPEG file with 75% quality.

The `IdentifyViewModel` now uses this `compressImage` function to process the selected image before creating the multipart request for the PlantNet API.

Signed-off-by: Abel 'Akronix' Serrano Juste <akronix5@gmail.com>