refactor: Remove legacy CameraScreen and related logic
This commit removes the deprecated `CameraScreen.kt` file, which relied on the `ACTION_IMAGE_CAPTURE` intent to take pictures. This component is no longer in use, as the application has migrated to a `CameraX`-based implementation for a more integrated experience.

As part of this cleanup, the unused `identifyOld()` function has also been removed from the `IdentifyViewModel`.

Signed-off-by: Abel 'Akronix' Serrano Juste <akronix5@gmail.com>